Create the separate controller for Goal Component
