+++ src/main/java/com/tvm/internal/edit/controller/GoalsController.java
@@ -1,0 +1,47 @@
+package com.tvm.internal.edit.controller;
+
+import com.tvm.internal.edit.model.Goal;
+import com.tvm.internal.edit.serviceImpl.GoalServiceImpl;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+import java.util.Optional;
+
+@RestController
+@RequestMapping
+public class GoalsController {
+
+    @Autowired
+    private GoalServiceImpl goalService;
+
+
+    @GetMapping("/goals")
+    public ResponseEntity<List<Goal>> getAllGoals() {
+        return ResponseEntity.ok(goalService.getAllGoals());
+    }
+
+    @GetMapping("/goals/{id}")
+    public ResponseEntity<Goal> getGoalById(@PathVariable Long id) {
+        Optional<Goal> goal = goalService.getGoalById(id);
+        return goal.map(ResponseEntity::ok)
+                .orElseGet(() -> ResponseEntity.notFound().build());
+    }
+
+    @PostMapping("/goals")
+    public ResponseEntity<Goal> createGoal(@RequestBody Goal goal) {
+        return ResponseEntity.ok(goalService.createGoal(goal));
+    }
+
+    @PutMapping("/goals/{id}")
+    public ResponseEntity<Goal> updateGoal(@PathVariable Long id, @RequestBody Goal goal) {
+        return ResponseEntity.ok(goalService.updateGoal(id, goal));
+    }
+
+    @DeleteMapping("/goals/{id}")
+    public ResponseEntity<Void> deleteGoal(@PathVariable Long id) {
+        goalService.deleteGoal(id);
+        return ResponseEntity.noContent().build();
+    }
+}
