+++ src/main/java/com/tvm/internal/edit/controller/TeamsController.java
@@ -19,9 +19,6 @@ public class TeamsController {
     @Autowired
     private TeamsServiceImpl teamService;
 
-    @Autowired
-    private GoalServiceImpl goalService;
-
     @GetMapping("/teams")
     public List<Teams> getAllTeams() {
         return teamService.getAllTeams();
@@ -45,33 +42,4 @@ public class TeamsController {
         return ResponseEntity.noContent().build();
     }
 
-
-    @GetMapping("/goal")
-    public ResponseEntity<List<Goal>> getAllGoals() {
-        return ResponseEntity.ok(goalService.getAllGoals());
-    }
-
-
-    @GetMapping("/goal/{id}")
-    public ResponseEntity<Goal> getGoalById(@PathVariable Long id) {
-        Optional<Goal> goal = goalService.getGoalById(id);
-        return goal.map(ResponseEntity::ok)
-                .orElseGet(() -> ResponseEntity.notFound().build());
-    }
-
-    @PostMapping("/goal")
-    public ResponseEntity<Goal> createGoal(@RequestBody Goal goal) {
-        return ResponseEntity.ok(goalService.createGoal(goal));
-    }
-
-    @PutMapping("/goal/{id}")
-    public ResponseEntity<Goal> updateGoal(@PathVariable Long id, @RequestBody Goal goal) {
-        return ResponseEntity.ok(goalService.updateGoal(id, goal));
-    }
-
-    @DeleteMapping("/goal/{id}")
-    public ResponseEntity<Void> deleteGoal(@PathVariable Long id) {
-        goalService.deleteGoal(id);
-        return ResponseEntity.noContent().build();
-    }
 }
